# services/api-gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway
  version: 1.0.0
  description: |
    Единая точка входа в интернет-магазин. Все клиентские запросы идут через
    API Gateway, который проверяет JWT-токен, подставляет заголовки
    `X-User-Id` / `X-User-Roles` и проксирует вызовы в микросервисы:

    | Раздел API        | Микросервис     | Транспорт |
    |-------------------|-----------------|-----------|
    | `/auth/*`         | auth-service    | HTTP      |
    | `/users/*`        | user-service    | HTTP      |
    | `/products/*`     | product-service | HTTP      |
    | `/categories/*`   | product-service | HTTP      |
    | `/inventory/*`    | inventory-service | HTTP    |
    | `/orders/*`       | order-service   | HTTP      |
    | остатки в каталоге | inventory-service | gRPC    |

    ## Как тестировать

    1. `POST /api/v1/auth/register` — регистрация (получите письмо с токеном подтверждения).
    2. `POST /api/v1/auth/login` — вход. В ответе придёт `access_token` и cookie `refresh_token`.
    3. Нажмите **Authorize** в Swagger UI и вставьте `access_token` (без слова `Bearer`).
    4. Все защищённые эндпоинты станут доступны.

    ## Авторизация

    - Защищённые эндпоинты требуют заголовок `Authorization: Bearer <access_token>`.
    - `access_token` — короткоживущий JWT (RS256, issuer `auth-service`).
    - `refresh_token` передаётся в HttpOnly-cookie `refresh_token`
      и используется только в `POST /api/v1/auth/refresh` и `POST /api/v1/auth/logout`.
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/inventory/{productId}:
    parameters:
      - $ref: '#/components/parameters/InventoryProductIdPath'

    get:
      summary: Остатки товара
      description: |
        Публичный эндпоинт. Возвращает доступное количество единиц
        товара на складе.

        Если для товара ещё не создана запись об остатках —
        возвращается `404`.
      tags: [inventory]
      operationId: getStock
      security: []
      responses:
        '200':
          description: Остатки товара
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
              example:
                product_id: 507f1f77bcf86cd799439011
                available_quantity: 12
                in_stock: true
        '404':
          description: Запись об остатках не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/inventory:
    post:
      summary: Создать запись об остатках
      description: |
        Создаёт остатки для товара с указанным начальным количеством.
        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.

        `product_id` — это `id` товара из каталога.
        Если запись уже существует — возвращается `409`.
      tags: [inventory]
      operationId: createStock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockRequest'
            examples:
              default:
                value:
                  product_id: 507f1f77bcf86cd799439011
                  initial_quantity: 100
      responses:
        '201':
          description: Остатки созданы
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Stock created successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Запись об остатках для товара уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/inventory/{productId}/add:
    parameters:
      - $ref: '#/components/parameters/InventoryProductIdPath'

    post:
      summary: Пополнить остатки
      description: |
        Увеличивает доступное количество товара на указанное значение
        (приход на склад). Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
      tags: [inventory]
      operationId: addStock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStockRequest'
            examples:
              default:
                value:
                  quantity: 25
      responses:
        '200':
          description: Остатки пополнены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Stock updated successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Запись об остатках не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  # ────────────────────────────── Orders ──────────────────────────────
...
  /api/v1/orders/{id}/payment/success:
    parameters:
      - $ref: '#/components/parameters/OrderIdPath'

    post:
      summary: Подтверждение успешной оплаты
      description: |
        Обрабатывает событие об успешной оплате: переводит заказ
        в статус `PAID`. Идемпотентна по `event_id` — повторная доставка
        того же события не меняет состояние заказа повторно.

        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
      tags: [orders]
      operationId: paymentSuccess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentSuccessRequest'
            examples:
              default:
                value:
                  event_id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  correlation_id: 1b671a64-40d5-491e-99b0-da01ff1f3341
      responses:
        '204':
          description: Оплата подтверждена
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/orders/{id}/payment/failed:
    parameters:
      - $ref: '#/components/parameters/OrderIdPath'

    post:
      summary: Сообщение об отказе оплаты
      description: |
        Обрабатывает событие об отказе оплаты: отменяет заказ и
        возвращает зарезервированные остатки на склад.
        Идемпотентна по `event_id`.

        Требуется `ROLE_MANAGER` или `ROLE_ADMIN`.
      tags: [orders]
      operationId: paymentFailed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentFailedRequest'
            examples:
              default:
                value:
                  event_id: 7c9e6679-7425-40de-944b-e07fc1f90ae7
                  correlation_id: 1b671a64-40d5-491e-99b0-da01ff1f3341
                  reason: insufficient funds
      responses:
        '204':
          description: Отказ по оплате обработан, заказ отменён
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Недопустимый переход статуса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access-токен, полученный в `POST /api/v1/auth/login`
        или `POST /api/v1/auth/refresh`.

        В Swagger UI нажмите **Authorize** и вставьте только значение
        токена, без префикса `Bearer `.

        Получить токен:
        ```bash
        curl -s -X POST http://localhost:4000/api/v1/auth/login \
          -H 'Content-Type: application/json' \
          -d '{"email":"user@example.com","password":"password123"}'
        ```

        Токен подписан RS256, issuer — `auth-service`. Gateway проверяет
        подпись, срок действия и `exp`/`nbf` с небольшим допуском (leeway).
        Истёкший или подделанный токен → `401`.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: |
        HttpOnly-cookie с refresh-токеном. Выставляется автоматически
        ответами `login`, `oauth/callback` и `refresh`.

        Cookie имеет флаги `Secure` и `Path=/api/v1/auth/refresh`,
        поэтому по `http://localhost:4000` браузер её не сохранит —
        для локального тестирования задайте значение cookie вручную
        через **Authorize** в Swagger UI.
  parameters:
...
    InventoryProductIdPath:
      name: productId
      in: path
      required: true
      description: ID товара, остатки которого запрашиваются (`id` из каталога)
      schema:
        type: string
      example: 507f1f77bcf86cd799439011
...
    OrderIdPath:
      name: id
      in: path
      required: true
      description: ID заказа (UUID)
      schema:
        type: string
        format: uuid
      example: 6ba7b810-9dad-11d1-80b4-00c04fd430c8
...
  schemas:
    ErrorResponse:
      type: object
      description: Единый формат ошибки для всех эндпоинтов шлюза
      required: [error]
      properties:
        error:
          type: string
          description: Сообщение об ошибке (может быть на русском языке)
          example: Проверьте корректность данных
...
    Stock:
      type: object
      description: Остатки товара на складе
      required: [product_id, available_quantity, in_stock]
      properties:
        product_id:
          type: string
          description: ID товара из каталога
          example: 507f1f77bcf86cd799439011
        available_quantity:
          type: integer
          description: Доступное количество единиц
          example: 12
        in_stock:
          type: boolean
          example: true
    CreateStockRequest:
      type: object
      required: [product_id]
      properties:
        product_id:
          type: string
          description: ID товара из каталога
          example: 507f1f77bcf86cd799439011
        initial_quantity:
          type: integer
          minimum: 0
          default: 0
          description: Начальное количество единиц товара
          example: 100
    AddStockRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          description: Количество единиц для прибавления к остатку
          example: 25
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Stock updated successfully
...
    PaymentSuccessRequest:
      type: object
      required: [event_id, correlation_id]
      properties:
        event_id:
          type: string
          format: uuid
          description: ID события оплаты — ключ идемпотентности обработки
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        correlation_id:
          type: string
          format: uuid
          description: ID корреляции для прослеживания по логам и трейсам
          example: 1b671a64-40d5-491e-99b0-da01ff1f3341
    PaymentFailedRequest:
      type: object
      required: [event_id, correlation_id, reason]
      properties:
        event_id:
          type: string
          format: uuid
          description: ID события оплаты — ключ идемпотентности обработки
          example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
        correlation_id:
          type: string
          format: uuid
          description: ID корреляции для прослеживания по логам и трейсам
          example: 1b671a64-40d5-491e-99b0-da01ff1f3341
        reason:
          type: string
          description: Причина отказа (например, недостаточно средств)
          example: insufficient funds
  responses:
...
    ValidationError:
      description: Ошибка валидации тела запроса (`Проверьте корректность данных`)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Проверьте корректность данных
    Unauthorized:
      description: |
        Не передан `Authorization: Bearer <token>`, токен невалиден или истёк
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noToken:
              summary: Токен не передан
              value:
                error: missing access token
            invalid:
              summary: Токен невалиден или истёк
              value:
                error: invalid or expired access token
    Forbidden:
      description: Недостаточно прав — у пользователя нет требуемой роли
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: insufficient permissions
...
    TooManyRequests:
      description: |
        Превышен лимит частоты запросов. Повторите через
        `X-RateLimit-Retry-After` секунд.
      headers:
        X-RateLimit-Limit:
          description: Максимальное число запросов в окне
          schema:
            type: integer
          example: 60
        X-RateLimit-Remaining:
          description: Сколько запросов осталось
          schema:
            type: integer
          example: 0
        X-RateLimit-RetryAfter:
          description: Через сколько секунд можно повторить
          schema:
            type: integer
          example: 2
        X-RateLimit-Reset:
          description: Unix-время сброса окна
          schema:
            type: integer
            format: int64
          example: 1736940000
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: too many requests
    InternalError:
      description: Внутренняя ошибка шлюза или недоступность микросервиса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error